Fixed all of the issues in the targeting engine builder and made it generalize.

+++ src/main/java/com/zuehlke/pgadmissions/domain/WeightedRelation.java
@@ -1,13 +1,15 @@
 package com.zuehlke.pgadmissions.domain;
 
+import java.math.BigDecimal;
+
 public abstract class WeightedRelation {
 
     public abstract Integer getId();
 
     public abstract void setId(Integer id);
 
-    public abstract Integer getRelationStrength();
+    public abstract BigDecimal getRelationStrength();
 
-    public abstract void setRelationStrength(Integer relationStrength);
+    public abstract void setRelationStrength(BigDecimal relationStrength);
 
 }
